FillPatcher class (#2972)
This adds a class FillPatcher for filling fine level data. It's not as
general as the various FillPatch functions (e.g., FillPatchTwoLevels).
However, it can reduce the amount of communication data. Suppose we use
RK2 with subcycling and the refinement ratio is 2. For each step on
level 0, there are two steps on level 1. With RK2, each fine step needs
to call FillPatch twice. So the total number of FillPatch calls is 4 in
the two fine steps. Using the free function, one ParallelCopy per
FillPatch call is needed for copying coarse data for spatial
interpolation. With the FillPatcher class, two ParallelCopy calls will
be done to copy old and new coarse data. Then these data will be used in
the four FillPatcher::fill calls. This new approach saves two
ParallelCopy calls per coarse step for a two levels run. It could save
more if the time stepping requires more substeps or the refinement ratio
is higher. Note that many of our AMReX codes use a time stepping
algorithm that needs only one FillPatch call per step. For those codes,
this new approach will not save any communication for a refinement ratio
of 2. However, it will save communication when the refinement ratio is
4.