MFIter::Finalize
Add a Finalize function to MFIter.

The idea about this is, that we can call this already before
destruction in Python, where `for` loops do not create
scope.

This function must be robust enough to be called again in
the constructor (or we need to add an extra bool to guard
that it is not called again in the destructor).