Runge-Kutta support for AMR (#2974)
This adds RK2, RK3 and RK4 in a new namespace RungeKutta. Together with
the enhanced FillPatcher class, these functions can be used for RK time
stepping in AMR simulations. A new function AmrLevel::RK is added for
AmrLevel based codes. See CNS::advance in Tests/GPU/CNS/CNS_advance.cpp
for an example of using the new AmrLevel::RK function.

The main motivation for this PR is that ghost cell filling for high
order (> 2) RK methods at coarse/fine boundary is non-trivial when there
is subcycling.

Co-authored-by: Jean M. Sexton <jmsexton@lbl.gov>