Fix a bug in ReduceOps when there is no work
This fixes a bug in the GPU version of ReduceOps.  If there is no work for
GPU, the tuple on the device is uninitialized. It is then copied to the host
and returned.  This could happen in ParReduce when a MultiFab has no boxes
on a process, as reported in #3000 by @BenWibking.  The bug was introduced
in April 2021.