Fix Multi-Level Open BC Solver
In the open BC solver, the enlarged level 0 grids are shifted so that the
lower left corner is at (0,0,0).  This allows the grids to be coarsened more
than if they are not shifted.  However, we did not shift the fine level
grids and this was a bug.  This bug is fixed here.