MLMG: template (#3035)
MLMG and various linear operator classes use MultiFab as their data
container.  The limitations are (1) it cannot do single precision when
amrex::Real is double and (2) it cannot handle data that need to be
stored
in multiple MultiFabs such as vectors on cell edges.  Recently, we have
experimented with amrex::Any that can be a type different from MultiFab.
However, this involves a lot of complicated and hacky changes. In this
PR,
we are trying to use template to make the interface more flexible.
Classes
MLMG, MLPoisson, MLALaplacian, MLABecLaplacian, MLABecLap,
MLCellLinop, MLLinOp, and MLCGSolver have been changed to templates.
More changes will follow.
    
For most users, this should not require any changes in their codes.
However, if one has classes derived from the linear operators in MLMG,
some
small changes are likely needed.  For example, `getNComp()` needs to be
`this->getNComp()` due to C++ name lookup rules.