Gpu Streams: Performance Tuning (#3058)
* Change the number of GPU streams from 4 to 2. Recent tests on summit
and crusher show that using 2 GPU streams seems to be better than 4 or 1
in most cases. One reason is that it reduces the cost associated with
stream synchronization, which could be significant for small kernels.

* In ~MFIter and ~StreamIter, there is no need to synchronize the
streams that were not used. It appears that for HIP there is very little
cost to sychronize a stream that does not have any GPU kerenels since
the last synchronization, but that's not the case for CUDA.

* Remove the dedicated stream for kernels outside MFIter, and simply use
the first stream used by MFIter.

* Remove synchronization in FabArray::define. It was needed when we were
putting FAB in managed memory, but that's no longer the case.