amrex::fillAsync (#3076)
This new function can be used to fill the elements in a vector type
container (e.g., Gpu::DeviceVector). If the element type is a struct
with several arithmetic types (e.g., GpuArray<Real,10>), the usual
ParallelFor does not have good performance because of the memory access
pattern. The new fillAsync function will use shared memory in that case
to improve the performance.