Improve AmrMesh::ChopGrids (#3081)
This tries to address the following issues.

(1) Suppose that the 2d domain has 40x8 cells, the blocking factor is 2
and the max grid size is 1024. Here the max grid size is set to a big
number so that we can have boxes as big as possible. If we have 4 MPI
processes, the old ChopGrids will chop the domain in x-direction into 5
parts. This is far from optimal.

(2) The issue in the example in (1) could be fixed by making the initial
chunk size to be the min of max grid size and domain size. This would
chop the domain into 4 boxes of 20x4 cells. However, this is still not
ideal because boxes close to squares should be preferred to long skinny
boxes. The new ChopGrids will chop the domain inot 4 boxes of 10x8
cells.