New knapsack function to reduce the cost of data movement in load balancing (#3079)
This adds a new function DistributionMapping::KnapSackProcessorMap that
makes a new DistributionMapping given an old DistributionMapping, a
vector of weights and a parameter affecting the amount of the data
movement that would be needed for switching from the old to the new
DistributionMapping.

In this new knapsack function, some boxes will be kept on the current
processes and the knapsack algorithm is applied to the rest of the
boxes. This may generate a less optimal distribution. However, the new
DistributionMapping is more similar to the old one, thus reducing the
cost of data movement during load balancing.