ParmParse: support nan and inf
When the ParamParse table is dumped, if a floating point number is a NaN, it
will be printed as nan.  If it's infinity, it will be printed as inf.
However, if one tried to read it back, the code will crash because ParmParse
did not recognize nan and inf as valid floating point numbers.  This commit
adds support for nan and inf in ParmParse.