Check component and derive names for `canDerive` and `get` (#3109)
`canDerive` and `get` in the `DeriveList` class only check the derive
name but not the variable component names. This code adds a check on the
variable names. @esclapez did this internally for PeleLMeX and we need
it for PeleC as well so I thought this should just belong in AMReX.