Fix bug in Device::Finalize: Forgot to clear streams
We forgot to clear the non-owning gpu_stream vector.  They still held the
old streams after we destroyed the streams in Device::Finalize.  This bug
affects codes that call amrex::Initialize and amrex::Finalize repeatedly.

Note that we have the non-owning gpu stream vector for OMP CPU thread
safety.

The bug was introduced in #1064.