CMake: Parallel HDF5 Detection
It turns out the words in the settings summary of HDF5 came from the words
used when HDF5 was configured with CMake.  The issue is `h5pcc -showconfig`
may show `Parallel HDF5: ON`, but the current versions of CMake uses
`Parallel HDF5: yes` to detect Parallel HDF5.  This commit adds a more
reliable way of detecting Parallel HDF5.