Allow choice of Intel compiler variant
This allows the user to specify `COMP = intel-classic` or `COMP =
intel-llvm` if they wish to use a specific one. Note that the existing
behaviour is maintained if `COMP = intel` (i.e. intel-llvm is chosen if
the `icpx` command exists and intel-classic is chosen otherwise).
However, the `AMREX_FCOMP`, `AMREX_CCOMP` and `lowercase_comp` makefile
variables are changed to one of `intel-llvm` or `intel-classic`
depending on which is used. This further allows the user to use
information about the used variant in their `Make.local` (for example
they may wish to set something like
```
CXX = mpiicpc -cxx=icpx
```
in the case `AMREX_CCOMP = intel-llvm`, `USE_MPI = TRUE` and they are
using Intel MPI.