CMake: Parallel HDF5 Detection (#3170)
## Summary

It turns out the words in the settings summary of HDF5 came from the
words used when HDF5 was configured with CMake. The issue is `h5pcc
-showconfig` may show `Parallel HDF5: ON`, but the current versions of
CMake uses `Parallel HDF5: yes` to detect Parallel HDF5. This commit
adds a more reliable way of detecting Parallel HDF5.

## Additional background

- https://github.com/spack/spack/issues/35546
- https://github.com/HDFGroup/hdf5/issues/2488
- https://gitlab.kitware.com/cmake/cmake/-/merge_requests/8234

## Checklist

The proposed changes:
- [x] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate