GNU Make: Allow choice of Intel compiler variant (#3175)
## Summary
This allows the user to specify `COMP = intel-classic` or `COMP =
intel-llvm` if they wish to use a specific one.

## Additional background
Note that the existing behaviour is maintained if `COMP = intel` (i.e.
intel-llvm is chosen if the `icpx` command exists and intel-classic is
chosen otherwise following #3126). However, the `AMREX_FCOMP`,
`AMREX_CCOMP` and `lowercase_comp` makefile variables are changed to one
of `intel-llvm` or `intel-classic` depending on which is used. This
further allows the user to use information about the used variant in
their `Make.local`. For example they may wish to set something like
```
CXX = mpiicpc -cxx=icpx
```
in the case `AMREX_CCOMP = intel-llvm`, `USE_MPI = TRUE` and they are
using Intel MPI.

## Checklist

The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [x] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [x] include documentation in the code and/or rst files, if appropriate