Update SUNDIALS NVector constructors (#3165)
## Summary

Update MultiFab NVector wrappers to accept a sundials context object.

## Additional background

Profiling (and in the future error handling) in sundials requires
objects to be create with a non-NULL context. This PR updates the
MultiFab NVector wrappers with an optional input for the sundials
context that defaults to the context created at initialization for the
corresponding OpenMP thread.