Fix issues in ParticleContainerBase (#3208)
* The move constructor and assignment operator were not safe, because
the m_gdb member could be the address of m_gdb_object. After the move
that can no longer be true. This is fixed by making m_gdb_object a
unique_ptr. Moving a unique_ptr does not change its pointer inside.

* Fix SetParticleBoxArray, SetParticleDistributionMapping and
SetParticleGeometry. The issue was these functions took a reference,
which could be aliasing what was inside the m_gdb_object. ParGDB has a
move assignment operator that could render the reference pointing to an
invalid object. So now these functions take by value. This was not an
issue before #3200, because the move assignment of ParGDB did copy
instead of move.