Warning about CPU core oversubscription
If the user forgets to set OMP_NUM_THREADS, OMP might be oversubscribing CPU
cores. If that's the case (to the best of our knowledge) and verbosity is
on, we print out a warning.

We use std::thread::hardware_concurrency to detect the number of cores
available. Note that this is only a hint according to the C++ standard. To
detect the number of MPI ranks per node, we use OMP_COMM_TYPE_NODE for Open
MPI and MPI_COMM_TYPE_SHARED for others.  Since MPI_COMM_TYPE_SHARED might
group processes across nodes, the return value of
ParallelDescriptor::NProcsPerNode might also be wrong.  Nevertheless, both
std::thread::hardware_concurrency and ParallelDescriptor::NProcsPerNode seem
to work well.