GitHub Cache Management
The total size of all caches in a GitHub repo is limited to 10 GB. So we
need to proactively manage the caches to avoid undesired cache eviction.

GitHub cache has its scope associated with branch. When a new PR is opened,
the last used cache of the default branch (i.e., development for AMReX) is
restored for the PR's actions. After the actions finish, the caches will be
saved under a unique name every time and only that PR has access to
it. (Presumably this is for security reason to avoid cache poison.) If the
actions are run many times, a large number of caches will be
accumulated. This uses a lot of space, even though we only need to keep the
most recent cache. In this PR, we implement the following approach to keep
only the last used cache for each job and remove all caches associated with
a PR when it's closed.

We have added a workflow that is triggered by the events of completed
workflows (e.g., LinuxClang workflow that has a number of jobs using cache.)
This new workflow runs on the default branch and it's therefore safe to be
granted write permission. Each triggering workflow saves the PR number as an
artifact, if it's a pull request run. This is a way to pass the PR number to
the cache cleanup workflow, which will clean up old caches associated with
the jobs in the triggering workflow. When a PR is closed or merged, a post
PR workflow is trigger, and it will in turn trigger another workflow that
cleans up all the caches associated with the PR. After the merge, the
actions will be run on the default branch. The newly added workflows will
also clean up the old caches in the default branch.