Notes on Cache Cleanup (#3237)
Add notes on cache cleanup workflows. This could be useful for debugging
and reasoning about how things work.

Fix a minor issue that only affects forks. To clean up cache left by
push events for a feature branch on a fork, we should use
github.event.workflow_run.head_branch (i.e., the feature branch) instead
of github.ref (i.e., the default branch). This is because the cache is
associated with the feature branch, whereas the cleanup workflow is run
on the default branch. This is not a problem for the main repo because a
direct push is not allowed. In the case of "indrect" push due to a PR
merge, the head branch is the default branch for us because we have only
one branch.