CMake Tests Setup
Previously most of the tests even if compiled with MPI were run on only one
process because they did not specify the number of tasks explicitly in their
setup. For the three tests that did specify, they were run twice during CI,
first on a single process and then on two processes, due to the setup in
Tests/CMakeLists.txt. The setup was similar for OpenMP.

This has changed. For tests compiled with OMP, they are run with two OMP
threads, otherwise they are run with two MPI processes if they are built
with MPI. This should make the CI tests faster.

Use 3 cores for macOS builds because GitHub provides 3 cores for macOS
runners.