ParticleTileData: No Restrict in Storage
Remove the `restrict` qualifier from `GpuArray` members
in `ParticleTileData`. We see compilation problems with, among
others Clang 14.0.6 and AppleClang, for this construct.

We expect that using `restrict` (`AMREX_RESTRICT`) on APIs as
well as aliased pointers in these arrays before access should
be sufficient to prevent that the compiler thinks that individual
arrays alias each other in hot loops.