CMake: Enable AMReX_PARTICLES by default (#3358)
This adds very little time to building the amrex library because most of
the particle codes are in headers. Note that `configure` using GNU Make
has had particle on by default already. Also note that this does not
affect amrex built with spack because it explicitly sets to the
particles variant to false.