CMake: AMReX_TEST_TYPE (#3357)
Add a new option, AMReX_TEST_TYPE=[All|Small]. Be default, all tests are
built when AMReX_ENABLE_TEST is on. When the type is small, only four
tests are enabled. If a CI job takes too long to compile and run all the
tests, we could switch to the small test type.

Re-enable tests and change the test type to Small for the nvhpc CI. We
had to disable it due to space limit in. It should no longer be a
problem for the small set of tests.