Amr Class: Synchronization of StateData's Time (#3375)
The synchronization of StateData's time is done in a virtual function,
AmrLevel::postCoarseTime. However, the issue is the override version of
the virtual function often forget to do it. In this PR, the job is
instead done in the Amr class directly. This should make the code more
robust.

Close #3374