Update Git workflow in CONTRIBUTING.md
Use `origin` as the name of the official amrex repo and `<Jane>` as the name
of a developer's fork. This is less confusing than using names like
`upstream` for the official repo and `origin` for the fork.

Remove the git rebase instruction, which is arguably out of the scope of
CONTRIBUTING.md.