New feature in EB: Add regular coarse levels (#3414)
This feature could be useful when the EB is so thin that it is only
resolved at the finest level and the usual way of generate coarse levels
fails with multi-cut or multi-value cells. When the EB is fully refined
to the finest level and there is no subcycling, no EB information is
really needed at coarse levels.

amrex::TagCutCells has been updated to work with the new feature.