Make ReduceData::value safer
Although ReduceData::value is meant to be called only once, the user might
call it multiple times. This will result in the wrong result for sum
reduction because of how the value function is implemented. In this PR, we
fix this defect by allowing the function be called multiple times.