generalize interpolaters to not need or use extra values in a directi… (#3415)
…on for which the refinement ratio is 1

The previous version of the interpolation routines in AmrCore would grow
the "CoarseBox" in all directions even if one of the directions had
refinement ratio 1. This PR changes that behavior so the CoarseBox is
not grown, and slopes are not computed, in any direction for which the
refinement ratio is 1.