Memory Efficient AoS Particle Sorting (#3427)
Because of the size of AoS particle, creating a temporary vector for all
the AoS data during reording might use too much memory. In fact, this is
usually the memory usage bottleneck in WarpX simulations. In this new
approach, we divide the AoS into smaller chunks and reorder them one at
a time, thus significantly reducing the peak memory usage.

Thanks @AlexanderSinn for the suggestion.