Re-Enable FPE Trap for SYCL (#3430)
FPE trapping was disabled for SYCL in the past due to a bug in the Intel
graphics compiler. The bug has long been fixed.

Also remove the macro for PGI, because we don't need to support PGI < 16
anymore.

More information about the bug. The JIT compiler computed `sqrt(-1)` and
was caught by AMReX's signal handler.