ParmParse: Line Continuation
In #3433, we added an option to disable multi-line support to avoid certain
types of errors. Since multi-line support seems to be desirable feature, we
add a new feature that uses `\` as line continuation. To avoid confusion, we
have decided to stop supporting the old style of multi-line values and make
it a runtime error. Hopefully it does not cause too much trouble for users.