CArena: Implement alloc_in_place (#3426)
First we try to see if CArena has free space that can be used to extend
the given pointer to the specified size range. If that's unsuccessful,
we use the normal alloc function to allocate memory.

If in-place allocation is successful during vector resizing, we can
avoid the cost of copying data to the new allocation address.