FillRandom & FillRandomNormal
Add new functions that fill a MultiFab or a 1d vector with random numbers
from a uniform or normal distribution.

The functions use CUDA and HIP's host APIs for random number
generations. Compared with our existing capability of using device APIs, the
new functions are slightly (about 15%) faster on Nvidia GPUs, whereas they
are 4x faster on the AMD GPU we tested. For Intel GPUs, we are still using
the device APIs, because we have not figured out how to use its host
APIs. More specially, we are missing some LDFLAGS at link time.

The functions work for MultiFabs with any index types (e.g.,
nodal). However, if the numbers on shared nodes need to be consistent across
boxes and MPI processes, it's the user's responsibility to call
OverrideSync. (This gives the user a chance to cache the mask used by
OverrideSync.) It's also the user's responsibility to fix data consistency
between overlapping ghost and valid points, if that is needed.