fix bug in StateRedist and add multi-level functionality for FluxRedi… (#3491)
…st and StateRedist

This PR has two purposes. The first is to fix a bug in
StateRedistribution which occurred when the EB was near a fine-fine
boundary. In this case we weren't calculating certain arrays in a wide
enough band of ghost cells. This has been fixed by enlarging two of the
arrays and calculating them on the correct number of ghost cells.

The second aspect of this PR is to make available the multi-level
StateRedistribution and FluxRedistribution algorithms. Using these will
require application codes to make changes in the calling routines. This
PR does not break the way application codes currently call the
single-level routines.