Update gpu elixir and async array synchronization for sycl (#3498)
Updated the way the GpuElixir and GpuAsyncArray synchronize. Instead of
a Gpu::streamSynchronize(); it now uses a host task by default since it
was implemented in SYCL 2020. Removed the codeplay-extension version.

In addition, the option SYCL_PARALLEL_LINK_JOBS was added to dpcpp.mak.
Setting it to a positive integer <n> then allows linking the SYCL code
with <n> threads, speeding up the whole compilation process, especially
when using AOT in SYCL.